changed float to double

+++ naumov.andrei/A1/base-types.hpp
@@ -3,14 +3,14 @@
 
 struct point_t
 {
-  float x;
-  float y;
+  double x;
+  double y;
 };
 
 struct rectangle_t
 {
-  float width;
-  float height;
+  double width;
+  double height;
   point_t pos;
 };
